WEB: Use OO for objects
