+++ include/Objects/Document.php
@@ -7,31 +7,17 @@ namespace ScummVM\Objects;
 class Document extends BasicObject
 {
     private $url;
-    private $description;
 
     /* Document object constructor. */
     public function __construct($data)
     {
-        $this->name = $data['name'];
+        parent::__construct($data);
         $this->url = $data['url'];
-        $this->description = $data['description'];
-    }
-
-    /* Get the name. */
-    public function getName()
-    {
-        return $this->name;
     }
 
     /* Get the URL. */
     public function getURL()
     {
         return $this->url;
-    }
-
-    /* Get the description. */
-    public function getDescription()
-    {
-        return $this->description;
     }
 }
